move controls to tardis package and added an AutoLandControl temporarily in the place of LandTypeControl cus its fun
