+++ src/main/java/mdteam/ait/core/blockentities/ConsoleBlockEntity.java
@@ -6,7 +6,7 @@ import mdteam.ait.core.AITBlockEntityTypes;
 import mdteam.ait.core.AITEntityTypes;
 import mdteam.ait.core.blocks.types.HorizontalDirectionalBlock;
 import mdteam.ait.core.entities.ConsoleControlEntity;
-import mdteam.ait.tardis.ControlTypes;
+import mdteam.ait.tardis.control.ControlTypes;
 import mdteam.ait.core.helper.TardisUtil;
 import mdteam.ait.data.AbsoluteBlockPos;
 import net.minecraft.block.BlockState;
