+++ src/main/java/mdteam/ait/core/blockentities/DoorBlockEntity.java
@@ -2,9 +2,7 @@ package mdteam.ait.core.blockentities;
 
 import mdteam.ait.api.tardis.ILinkable;
 import mdteam.ait.core.AITBlockEntityTypes;
-import mdteam.ait.core.AITItems;
 import mdteam.ait.core.blocks.types.HorizontalDirectionalBlock;
-import mdteam.ait.core.entities.control.impl.DoorControl;
 import mdteam.ait.core.helper.TardisUtil;
 import mdteam.ait.core.item.KeyItem;
 import mdteam.ait.data.AbsoluteBlockPos;
@@ -28,8 +26,6 @@ import mdteam.ait.tardis.TardisDesktop;
 import mdteam.ait.tardis.TardisManager;
 
 import java.util.Objects;
-
-import static mdteam.ait.AITMod.INTERIORDOORNBT;
 
 public class DoorBlockEntity extends BlockEntity implements ILinkable {
