+++ src/main/java/mdteam/ait/tardis/TardisConsole.java
@@ -1,6 +1,7 @@
 package mdteam.ait.tardis;
 
 import mdteam.ait.client.renderers.consoles.ConsoleEnum;
+import mdteam.ait.tardis.control.ControlTypes;
 
 public class TardisConsole {
