+++ src/main/java/mdteam/ait/tardis/control/impl/HandBrakeControl.java
@@ -1,0 +1,25 @@
+package mdteam.ait.tardis.control.impl;
+
+import mdteam.ait.tardis.control.Control;
+import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.client.world.ClientWorld;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
+import mdteam.ait.tardis.Tardis;
+
+public class HandBrakeControl extends Control {
+    public HandBrakeControl() {
+        super("handbrake");
+    }
+
+    // todo
+    @Override
+    public boolean runClient(Tardis tardis, ClientPlayerEntity player, ClientWorld world) {
+        return false;
+    }
+
+    @Override
+    public boolean runServer(Tardis tardis, ServerPlayerEntity player, ServerWorld world) {
+        return false;
+    }
+}
