+++ src/main/java/mdteam/ait/tardis/control/impl/RefuelerControl.java
@@ -1,0 +1,25 @@
+package mdteam.ait.tardis.control.impl;
+
+import mdteam.ait.tardis.control.Control;
+import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.client.world.ClientWorld;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
+import mdteam.ait.tardis.Tardis;
+
+public class RefuelerControl extends Control {
+    public RefuelerControl() {
+        super("refueler");
+    }
+
+    // todo
+    @Override
+    public boolean runClient(Tardis tardis, ClientPlayerEntity player, ClientWorld world) {
+        return false;
+    }
+
+    @Override
+    public boolean runServer(Tardis tardis, ServerPlayerEntity player, ServerWorld world) {
+        return false;
+    }
+}
