+++ src/main/java/mdteam/ait/tardis/control/impl/ThrottleControl.java
@@ -1,15 +1,13 @@
-package mdteam.ait.core.entities.control.impl;
+package mdteam.ait.tardis.control.impl;
 
-import mdteam.ait.core.AITSounds;
-import mdteam.ait.core.entities.control.Control;
-import mdteam.ait.core.helper.TardisUtil;
+import mdteam.ait.tardis.control.Control;
+import mdteam.ait.tardis.handler.PropertiesHandler;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import mdteam.ait.tardis.Tardis;
 import mdteam.ait.tardis.TardisTravel;
-import net.minecraft.sound.SoundCategory;
 
 public class ThrottleControl extends Control {
     public ThrottleControl() {
@@ -27,7 +25,7 @@ public class ThrottleControl extends Control {
         TardisTravel travel = tardis.getTravel();
 
         if (travel.getState() == TardisTravel.State.LANDED) {
-            travel.dematerialise(false);
+            travel.dematerialise(PropertiesHandler.willAutoLand(tardis.getProperties()));
         } else if (travel.getState() == TardisTravel.State.FLIGHT) {
             travel.materialise();
         }
