+++ src/main/java/mdteam/ait/tardis/control/impl/VisualiserControl.java
@@ -1,6 +1,6 @@
-package mdteam.ait.core.entities.control.impl;
+package mdteam.ait.tardis.control.impl;
 
-import mdteam.ait.core.entities.control.Control;
+import mdteam.ait.tardis.control.Control;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.server.network.ServerPlayerEntity;
