+++ src/main/java/mdteam/ait/tardis/control/impl/pos/YControl.java
@@ -1,4 +1,4 @@
-package mdteam.ait.core.entities.control.impl.pos;
+package mdteam.ait.tardis.control.impl.pos;
 
 public class YControl extends PosControl {
     public YControl() {
