+++ src/main/java/mdteam/ait/tardis/control/impl/pos/ZControl.java
@@ -1,4 +1,4 @@
-package mdteam.ait.core.entities.control.impl.pos;
+package mdteam.ait.tardis.control.impl.pos;
 
 public class ZControl extends PosControl {
     public ZControl() {
