+++ src/main/java/mdteam/ait/tardis/wrapper/server/ServerTardisConsole.java
@@ -1,7 +1,7 @@
 package mdteam.ait.tardis.wrapper.server;
 
 import mdteam.ait.client.renderers.consoles.ConsoleEnum;
-import mdteam.ait.tardis.ControlTypes;
+import mdteam.ait.tardis.control.ControlTypes;
 import mdteam.ait.tardis.ServerTardisManager;
 import mdteam.ait.tardis.Tardis;
 import mdteam.ait.tardis.TardisConsole;
